Adds Nano support for adding users to groups
Changes the logic of the current method so that it will
work as expected on a NanoServer instance.
At the moment the method shows an unusual behaviour
on such instances, by throwing unexpected OSErrors,
caused by trying to get 'addressof' of the given user.
Changing to the 'win32net' module and letting it do the
memory management when adding users to groups solves this issue.

Change-Id: I971b60d1a84c1250a8e8d2f82351324faabff836
Closes-Bug: #1643915