Implemented automatic updates plugin
Adds a plugin to configure Windows automatic updates. The automatic updates
configuration is retrieved from the metadata service and there is also a
configurable option to enabled or disable them.
In case that the configurable option is changed, and the metadata
service option is also present, the configurable option will apply instead of the
option retrieavable from the metadata.

Implements: blueprint windows-automated-updates-plugin
Co-Authored-By: Stefan Caraiman <scaraiman@cloudbasesolutions.com>
Co-Authored-By: Adrian Vladu <avladu@cloudbasesolutions.com>
Change-Id: Ida07a3a279321fe434a82a5123338679ec4506df